Refactor ambiguity solvers into object for better readability
