+++ core/src/main/kotlin/org/evomaster/core/output/naming/NamingHelperNumberedTestCaseNamingStrategy.kt
@@ -3,7 +3,6 @@ package org.evomaster.core.output.naming
 import org.evomaster.core.output.NamingHelper
 import org.evomaster.core.search.EvaluatedIndividual
 import org.evomaster.core.search.Solution
-import org.evomaster.core.search.action.Action
 
 class NamingHelperNumberedTestCaseNamingStrategy(
     solution: Solution<*>
@@ -11,7 +10,7 @@ class NamingHelperNumberedTestCaseNamingStrategy(
 
     private val namingHelper = NamingHelper()
 
-    override fun expandName(individual: EvaluatedIndividual<*>, nameTokens: MutableList<String>, ambiguitySolvers: List<(Action) -> List<String>>): String {
+    override fun expandName(individual: EvaluatedIndividual<*>, nameTokens: MutableList<String>, ambiguitySolvers: List<AmbiguitySolver>): String {
         return namingHelper.suggestName(individual)
     }
 }
