+++ core/src/main/kotlin/org/evomaster/core/output/naming/TestCaseNamingStrategy.kt
@@ -3,7 +3,6 @@ package org.evomaster.core.output.naming
 import org.evomaster.core.output.TestCase
 import org.evomaster.core.search.EvaluatedIndividual
 import org.evomaster.core.search.Solution
-import org.evomaster.core.search.action.Action
 
 /**
  * A naming strategy will help provide names to the generated test cases.
@@ -35,11 +34,11 @@ abstract class TestCaseNamingStrategy(
     /**
      * @param individual containing information for the test about to be named
      * @param nameTokens list to collect the identifiers which will be formatted into the test case name
-     * @param ambiguitySolvers list of functions receiving an action and returning a list of strings that will be added to the  test case name
+     * @param ambiguitySolvers list of solvers to disambiguate the test case names
      *
      * @return a String with extra information that will be included in the test name, regarding the EvaluatedIndividual
      */
-    protected abstract fun expandName(individual: EvaluatedIndividual<*>, nameTokens: MutableList<String>, ambiguitySolvers: List<(Action) -> List<String>> = emptyList()): String
+    protected abstract fun expandName(individual: EvaluatedIndividual<*>, nameTokens: MutableList<String>, ambiguitySolvers: List<AmbiguitySolver> = emptyList()): String
 
     /**
      * @param duplicatedIndividuals set containing the EvaluatedIndividuals sharing the same name
